Quiz App
Designed and implemented an Android Quiz App in Java and XML using Android Studio. This app is a part of Udacity/Google "Grow with Google Challenge Scholarship program project work.

The app has two subjects Math and Science depending on the selection of the subject, it will take you to the subject specific quiz page.